fix: contextmenu doesn't work on windows (#29)
On macOS, when you right click a link, it's text will be selected as well and therefore it's accessible through `link.selectionText`, but unfortunately it's not the case on windows. so when the extension tries to send a request to the API, it's text will be null and therefore nothing will be sent. Although `link.linkText` is available on firefox, so first check for it and if it didn't exist, then check for `link.selectionText` and if it didn't exist either, use the link url itself.

Fixes: #28